Add nested text formatting and loose separators
Now we can do things like bold AND italic text and the document
doesn't have to end in 2 new lines to parse successfully.

Moving forward:

Problem:

Document can't end with plain-text or a header. It says combinator
'many' has been applied to a parser that accepts an empty string.

TODO:

Fix it so the document can end with plain-text or a header or
nothing